Implement Animation System Prototype
Add Animation, Animator, and AnimationManager

Animator uses tween-based system to animate sprites based on information
provided in Animation assets, found in /resources/animations

Animations are files of type .reel

Animation system currently only loads frames into image cache and provides
name of current frame's image (upon request). Actual animation must be
handled in an animated variant of the SpriteRenderer Lua component.

Need to integrate SpriteRenderer and automate/streamline animation system.

+++ src/Actor.cpp
@@ -248,7 +248,7 @@ sol::table Actor::cppActorGetComponents(const std::string &type_name)
 {
     sol::table component_table = ComponentManager::GetLuaState()->create_table();
     
-    int components_found = 0;
+    uint16_t components_found = 0;
     
     for (auto &actor_component_pair : actor_components)
     {
